+++ lib/authority/authorizer.rb
@@ -27,9 +27,13 @@ module Authority
     Authority.adjectives.each do |adjective|
       class_eval <<-RUBY, __FILE__, __LINE__ + 1
         def self.#{adjective}_by?(user)
-          Authority.configuration.default_strategy.call(:#{adjective}, self, user)
+          default(:#{adjective}, user)
         end
       RUBY
+    end
+
+    def self.default(adjective, user)
+      Authority.configuration.default_strategy.call(adjective, self, user)
     end
 
   end